+++ src/components/navbar/AppDrawer.jsx
@@ -1,69 +1,28 @@
 import React from 'react';
 import clsx from 'clsx';
-import Drawer from '@material-ui/core/Drawer';
-import Hidden from '@material-ui/core/Hidden';
-import TopDivider from './Drawer';
 import EventCard from "../EventCard";
-import AppBarNav from "./AppBar";
-import {CssBaseline, makeStyles, useTheme, Divider} from "@material-ui/core";
+import AppBarNav from "./nav-components/AppBar";
+import {CssBaseline, makeStyles} from "@material-ui/core";
 const drawerWidth = 240;
 const useStyles = makeStyles((theme) => ({
   root: {
     display: 'flex',
   },
-  drawer: {
-    [theme.breakpoints.up('sm')]: {
-      width: drawerWidth,
-      flexShrink: 0,
-    },
-  },
   // necessary for content to be below app bar
   toolbar: theme.mixins.toolbar,
-  drawerPaper: {
-    width: drawerWidth,
-  },
+  drawerPaper: { width: drawerWidth},
   content: {
     flexGrow: 1,
     padding: theme.spacing(3),
+    marginLeft: drawerWidth,
   },
 }));
 function ResponsiveDrawer(props) {
   const classes = useStyles();
-  const theme = useTheme();
-  const [mobileOpen, setMobileOpen] = React.useState(false);
-  const handleDrawerToggle = () => {
-    setMobileOpen(!mobileOpen);
-  };
-  const drawer = (
-    <div>
-      <div className={classes.toolbar} />
-      <Divider />
-      <TopDivider />
-    </div>
-  );
   return (
     <div className={classes.root}>
       <CssBaseline />
       <AppBarNav />
-      <nav className={classes.drawer} aria-label="mailbox folders">
-        <Hidden smUp implementation="css">
-          <Drawer
-            variant="temporary"
-            anchor={theme.direction === 'rtl' ? 'right' : 'left'}
-            open={mobileOpen}
-            onClose={handleDrawerToggle}
-            classes={{ paper: classes.drawerPaper }}
-            ModalProps={{ keepMounted: true }}>
-            {drawer}
-          </Drawer>
-        </Hidden>
-        <Hidden xsDown implementation="css">
-          <Drawer classes={{ paper: classes.drawerPaper}}
-                  variant="permanent" open>
-            {drawer}
-          </Drawer>
-        </Hidden>
-      </nav>
       <main className={clsx(classes.content)}>
         <div className={classes.toolbar} />
         <EventCard/>
