+++ src/components/navbar/nav-components/DrawerFunction.jsx
@@ -1,0 +1,56 @@
+import React from 'react';
+import Drawer from '@material-ui/core/Drawer';
+import Hidden from '@material-ui/core/Hidden';
+import TopDivider from './DrawerList';
+import { makeStyles, useTheme, Divider} from "@material-ui/core";
+const drawerWidth = 240;
+const useStyles = makeStyles((theme) => ({
+  root: {
+    display: 'flex',
+  },
+  drawer: {
+    [theme.breakpoints.up('sm')]: {
+      width: drawerWidth,
+      flexShrink: 0,
+    },
+  },
+}));
+function DrawerFunction(props) {
+  const classes = useStyles();
+  const theme = useTheme();
+  const [mobileOpen, setMobileOpen] = React.useState(false);
+  const handleDrawerToggle = () => {
+    setMobileOpen(!mobileOpen);
+  };
+  const drawer = (
+    <div>
+      <div className={classes.toolbar} />
+      <Divider />
+      <TopDivider />
+    </div>
+  );
+  return (
+    <div className={classes.root}>
+      <nav className={classes.drawer} aria-label="mailbox folders">
+        <Hidden smUp implementation="css">
+          <Drawer
+            variant="temporary"
+            anchor={theme.direction === 'rtl' ? 'right' : 'left'}
+            open={mobileOpen}
+            onClose={handleDrawerToggle}
+            classes={{ paper: classes.drawerPaper }}
+            ModalProps={{ keepMounted: true }}>
+            {props.drawer}
+          </Drawer>
+        </Hidden>
+        <Hidden xsDown implementation="css">
+          <Drawer classes={{ paper: classes.drawerPaper}}
+                  variant="permanent" open>
+            {props.drawer}
+          </Drawer>
+        </Hidden>
+      </nav>
+    </div>
+  );
+}
+export default DrawerFunction;
