+++ src/components/navbar/nav-components/SearchBox.jsx
@@ -1,0 +1,61 @@
+import React from 'react';
+import SearchIcon from "@material-ui/icons/Search";
+import {makeStyles, fade, InputBase} from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+    root: {
+        display: 'flex',
+      },
+    search: {
+        position: 'relative',
+        borderRadius: theme.shape.borderRadius,
+        backgroundColor: fade(theme.palette.common.white, 0.15),
+        '&:hover': { backgroundColor: fade(theme.palette.common.white, 0.25)},
+        marginLeft: 0,
+        width: '100%',
+        [theme.breakpoints.up('sm')]: {
+          marginLeft: theme.spacing(1),
+          width: 'auto',
+        },
+    },
+    searchIcon: {
+      padding: theme.spacing(0, 2),
+      height: '100%',
+      position: 'absolute',
+      pointerEvents: 'none',
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    inputRoot: { color: 'inherit'},
+    inputInput: {
+      padding: theme.spacing(1, 1, 1, 0),
+      // vertical padding + font size from searchIcon
+      paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
+      transition: theme.transitions.create('width'),
+      width: '100%',
+      [theme.breakpoints.up('sm')]: {
+        width: '12ch',
+        '&:focus': { width: '20ch'},
+      },
+    },
+  }));
+
+  function SearchBoxNav() {
+    const classes = useStyles();
+    return (
+        <>
+           <div className={classes.shiftRight}>
+           <div className={classes.search}>
+           <div className={classes.searchIcon}>
+             <SearchIcon />
+           </div>
+           <InputBase
+             placeholder="Search…"
+             classes={{root: classes.inputRoot,input: classes.inputInput}}
+             inputProps={{ 'aria-label': 'search' }}/>
+           </div>
+           </div>
+        </>
+)};
+export default SearchBoxNav;
